Crate hacspec_secp256k1

An implementation of the secp256k1 curve in the hacspec specification language.

Many properties of this library have been proven in Hacspec_Secp256k1 (documentation).

Structs

FieldCanvas

ScalarCanvas

Secp256k1FieldElement

Secp256k1Scalar

Functions

GENERATOR

Returns the base point, G, for the Secp256k1 curve in affine coordinates.

INFINITY

Generates an affine representation of point at infinity (uses a placeholder off the curve).

add_points

Adds two arbitrary points in affine coordinates.

batch_scalar_multiplication

Calculates the sum a_1 * P_1 + … + a_i * P_i + … + A_m * P_m efficiently.

double_point

Doubles the given point in affine coordinates.

is_infinity

Checks whether the given point is the point at infinity.

is_point_on_curve

Checks if the given point is a valid point on the curve.

neg_point

Negates the given point in affine coordinates.

scalar_multiplication

Performs scalar multiplication on the given point in affine coordinates.

Type Definitions

Affine

Represents a point in affine coordinates. The point at infinity is encoded as the otherwise invalid value (1,0).